chore(server): moving into same folder
