+++ packages/twenty-server/test/integration/metadata/suites/field-metadata/enum/create-one-field-metadata-multi-select.integration-spec.ts
@@ -1,8 +1,8 @@
-import { UpdateCreateFieldMetadataSelectTestCase } from 'test/integration/metadata/suites/field-metadata/update-create-one-field-metadata-select-tests-cases';
+import { UpdateCreateFieldMetadataSelectTestCase } from 'test/integration/metadata/suites/field-metadata/enum/field-metadata-select-and-multi-select-common-tests-cases';
 import { createOneFieldMetadata } from 'test/integration/metadata/suites/field-metadata/utils/create-one-field-metadata.util';
 import {
-    LISTING_NAME_PLURAL,
-    LISTING_NAME_SINGULAR,
+  LISTING_NAME_PLURAL,
+  LISTING_NAME_SINGULAR,
 } from 'test/integration/metadata/suites/object-metadata/constants/test-object-names.constant';
 import { deleteOneObjectMetadata } from 'test/integration/metadata/suites/object-metadata/utils/delete-one-object-metadata.util';
 import { FieldMetadataType } from 'twenty-shared/types';
