+++ packages/twenty-server/test/integration/metadata/suites/field-metadata/enum/field-metadata-select-and-multi-select-common-tests-cases.ts
