+++ packages/twenty-server/test/integration/metadata/suites/field-metadata/enum/update-one-field-metadata-select.integration-spec.ts
@@ -1,7 +1,7 @@
 import {
   UPDATE_CREATE_ONE_FIELD_METADATA_SELECT_TEST_CASES,
   UpdateCreateFieldMetadataSelectTestCase,
-} from 'test/integration/metadata/suites/field-metadata/update-create-one-field-metadata-select-tests-cases';
+} from 'test/integration/metadata/suites/field-metadata/enum/field-metadata-select-and-multi-select-common-tests-cases';
 import { CreateOneFieldFactoryInput } from 'test/integration/metadata/suites/field-metadata/utils/create-one-field-metadata-query-factory.util';
 import { createOneFieldMetadata } from 'test/integration/metadata/suites/field-metadata/utils/create-one-field-metadata.util';
 import { updateOneFieldMetadata } from 'test/integration/metadata/suites/field-metadata/utils/update-one-field-metadata.util';
